abstract error messages
