+++ lib/errorMessages.js
@@ -1,0 +1,17 @@
+const errorMessages = {
+  noArticle: {
+    error: true,
+    message: 'Article does not exist',
+  },
+  noComment: {
+    error: true,
+    message: 'Comment does not exist'
+  },
+  noReply: {
+    error: true,
+    message: 'Reply does not exist'
+  }
+};
+
+
+export default errorMessages;
