+++ lib/helper.js
@@ -27,24 +27,9 @@ const throwErrorOnNonExistingUser = async (userId) => {
   if (!existingUser) throw new Error('Error: User doen\'t exist');
   return existingUser;
 };
-const errorMessages = {
-  noArticle: {
-    error: true,
-    message: 'Article does not exist',
-  },
-  noComment: {
-    error: true,
-    message: 'Comment does not exist'
-  },
-  noReply: {
-    error: true,
-    message: 'Reply does not exist'
-  }
-};
 
 export default {
   throwErrorOnBadRequest,
   throwErrorOnNonExistingUser,
-  findItem,
-  errorMessages
+  findItem
 };
